requirements.txt
Flask==3.1.2
PyJWT==2.11.0
python-dotenv==1.2.1
python_bcrypt==0.3.2
twelvedata==1.2.28
